refactor AddArgumentsAsVariables
